update+delete and conflict finished

+++ fin-wallet-api/src/main/java/com/codegym/finwallet/dto/payload/request/WalletRequest.java
@@ -11,7 +11,7 @@ import lombok.NoArgsConstructor;
 @Builder
 public class WalletRequest {
     private String icon;
-    private double amount;
+    private float amount;
     private String currentType;
     private String description;
     private String name;
